added common field base to simplify everything
