+++ Sources/Vein/ModelContext/Values/PredicateBuilder.swift
@@ -55,7 +55,7 @@ public struct PredicateBuilder<T: PersistentModel>: Sendable, Hashable, AnyPredi
                     .first(
                         where: { $0.instanceKey == key }
                     )!
-                    .wrappedValue as! V
+                    .persistableValue as! V
             )
             .fieldIsEqualTo(
                 sqliteValue
@@ -91,7 +91,7 @@ public struct PredicateBuilder<T: PersistentModel>: Sendable, Hashable, AnyPredi
                     .first(
                         where: { $0.instanceKey == key }
                     )!
-                    .wrappedValue as! V
+                    .persistableValue as! V
             )
             .fieldIsNotEqualTo(
                 sqliteValue
@@ -128,7 +128,7 @@ public struct PredicateBuilder<T: PersistentModel>: Sendable, Hashable, AnyPredi
                     .first(
                         where: { $0.instanceKey == key }
                     )!
-                    .wrappedValue as! V
+                    .persistableValue as! V
             )
             .fieldIsBiggerThan(
                 sqliteValue
@@ -165,7 +165,7 @@ public struct PredicateBuilder<T: PersistentModel>: Sendable, Hashable, AnyPredi
                     .first(
                         where: { $0.instanceKey == key }
                     )!
-                    .wrappedValue as! V
+                    .persistableValue as! V
             )
             .fieldIsSmallerThan(
                 sqliteValue
@@ -202,7 +202,7 @@ public struct PredicateBuilder<T: PersistentModel>: Sendable, Hashable, AnyPredi
                     .first(
                         where: { $0.instanceKey == key }
                     )!
-                    .wrappedValue as! V
+                    .persistableValue as! V
             )
             .fieldIsSmallerOrEqual(
                 sqliteValue
@@ -239,7 +239,7 @@ public struct PredicateBuilder<T: PersistentModel>: Sendable, Hashable, AnyPredi
                     .first(
                         where: { $0.instanceKey == key }
                     )!
-                    .wrappedValue as! V
+                    .persistableValue as! V
             )
             .fieldIsBiggerOrEqual(
                 sqliteValue
